Added a backend part for the website.
I added in Spring boot for the backend. This will allow me to use Java. I haven't got react and spring boot to communicate yet, but I made a simple test program to test if websites were up and running by making a simple connection and sending a GET request. Depending on the request code it will determine if the site is up or if it is down.